requirements.txt
asgiref==3.8.1
Django==5.0.4
djangorestframework==3.15.1
drf-yasg==1.21.7
environs==11.0.0
inflection==0.5.1
marshmallow==3.21.1
mysqlclient==2.2.4
packaging==24.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.5.0
typing_extensions==4.11.0
uritemplate==4.1.1
